backend-api/requirements.txt
fastapi
uvicorn
psycopg2-binary
qdrant-client
openai
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
qdrant-client
openai
langchain
langchain-community
langchain-openai
langchain-text-splitters
